ui. moving components and mixins to library
- TaskService, StorageService
