+++ core/ui/package.json
@@ -16,7 +16,7 @@
     "@carbon/icons-vue": "^10.31.0",
     "@carbon/themes": "^10.34.0",
     "@carbon/vue": "^2.38.2",
-    "andrelib": "0.0.18",
+    "andrelib": "0.0.20",
     "await-to-js": "^3.0.0",
     "axios": "^0.21.1",
     "carbon-components": "^10.34.0",
