+++ core/ui/src/App.vue
@@ -15,16 +15,14 @@ import ShellHeader from "./components/ShellHeader";
 import SideMenu from "./components/SideMenu";
 import MobileSideMenu from "./components/MobileSideMenu";
 import axios from "axios";
-import StorageService from "@/mixins/storage";
 import WebSocketService from "@/mixins/websocket";
 import { mapState } from "vuex";
 import { mapActions } from "vuex";
 import to from "await-to-js";
 import LoginService from "@/mixins/login";
 import TaskErrorModal from "@/components/TaskErrorModal";
-import TaskService from "@/mixins/task";
 import NotificationService from "@/mixins/notification";
-import { UtilService } from "andrelib"; ////
+import { UtilService, TaskService, StorageService } from "andrelib"; ////
 
 //// remove andrelib from package.json
 
@@ -51,10 +49,10 @@ export default {
     // register to events
     this.$root.$on("login", this.initNs8);
     this.$root.$on("logout", this.logout);
-    this.$root.$on("createNotificationForApp", this.createNotificationForApp);
+    // this.$root.$on("createNotificationForApp", this.createNotificationForApp); //// not used?
     this.$root.$on(
-      "createTaskErrorNotificationForApp",
-      this.createTaskErrorNotificationForApp
+      "createTaskErrorNotification",
+      this.createTaskErrorNotification
     );
 
     this.configureAxiosInterceptors();
@@ -127,7 +125,7 @@ export default {
       this.$root.$on(taskAction + "-completed", this.listUpdatesCompleted);
 
       const res = await to(
-        this.createTask({
+        this.createClusterTask({
           action: taskAction,
           extra: {
             title: this.$t("action." + taskAction),
@@ -165,13 +163,20 @@ export default {
 
       this.setUpdatesInStore(updates);
     },
-    createNotificationForApp(notification) {
-      // create a notification as requested by an external app
+    // createNotificationForApp(notification) { //// not used?
+    //   // create a notification as requested by an external app
+    //   this.createNotification(notification);
+    // },
+    createTaskErrorNotification(err, message) {
+      // create a task error notification as requested by an external app ////
+      // this.createTaskErrorNotification(err, message); ////
+
+      const notification = {
+        title: message,
+        description: this.getErrorMessage(err),
+        type: "error",
+      };
       this.createNotification(notification);
-    },
-    createTaskErrorNotificationForApp(err, message) {
-      // create a task error notification as requested by an external app
-      this.createTaskErrorNotification(err, message);
     },
   },
 };
