+++ core/ui/src/components/ShellHeader.vue
@@ -80,10 +80,10 @@ import { mapGetters } from "vuex";
 import GlobalSearch from "@/components/GlobalSearch";
 import AppDrawer from "@/components/AppDrawer";
 import Search20 from "@carbon/icons-vue/es/search/20";
-import StorageService from "@/mixins/storage";
 import LoginService from "@/mixins/login";
 import WebSocketService from "@/mixins/websocket";
 import NotificationDrawer from "@/components/NotificationDrawer";
+import { StorageService } from "andrelib"; ////
 
 export default {
   name: "ShellHeader",
