--- core/ui/src/mixins/task.js
@@ -1,127 +1,0 @@
-import StorageService from "@/mixins/storage";
-import ErrorFilled16 from "@carbon/icons-vue/es/error--filled/16";
-import CheckmarkFilled16 from "@carbon/icons-vue/es/checkmark--filled/16";
-import WarningFilled16 from "@carbon/icons-vue/es/warning--filled/16";
-import InformationFilled16 from "@carbon/icons-vue/es/information--filled/16";
-import { UtilService } from "andrelib"; ////
-import NotificationService from "@/mixins/notification";
-
-export default {
-  name: "TaskService",
-  mixins: [StorageService, UtilService, NotificationService],
-  methods: {
-    getTaskContext(taskPath) {
-      const token = this.getFromStorage("loginInfo")
-        ? this.getFromStorage("loginInfo").token
-        : "";
-      return this.axios.get(`${this.$root.apiUrl}/${taskPath}/context`, {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      });
-    },
-    getTaskStatus(taskPath) {
-      const token = this.getFromStorage("loginInfo")
-        ? this.getFromStorage("loginInfo").token
-        : "";
-      return this.axios.get(`${this.$root.apiUrl}/${taskPath}/status`, {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      });
-    },
-    getClusterTasks() {
-      const token = this.getFromStorage("loginInfo")
-        ? this.getFromStorage("loginInfo").token
-        : "";
-      return this.axios.get(`${this.$root.apiUrl}/tasks/cluster`, {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      });
-    },
-    createTask(taskData) {
-      const token = this.getFromStorage("loginInfo")
-        ? this.getFromStorage("loginInfo").token
-        : "";
-      return this.axios.post(`${this.$root.apiUrl}/cluster/tasks`, taskData, {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      });
-    },
-    getTaskTitle(task) {
-      if (
-        task &&
-        task.context &&
-        task.context.data &&
-        task.context.data.title
-      ) {
-        return task.context.data.title;
-      } else {
-        return "";
-      }
-    },
-    getTaskKind(task) {
-      switch (task.status) {
-        case "aborted":
-        case "validation-failed":
-          return "error";
-        case "completed":
-          return "success";
-        case "pending":
-          return "warning";
-        default:
-          return "info";
-      }
-    },
-    getTaskIcon(task) {
-      switch (task.status) {
-        case "aborted":
-        case "validation-failed":
-          return ErrorFilled16;
-        case "completed":
-          return CheckmarkFilled16;
-        case "pending":
-          return WarningFilled16;
-        default:
-          return InformationFilled16;
-      }
-    },
-    getTaskStatusDescription(task, rootTask = true) {
-      const taskAction = task.context.action;
-      const taskOrSubtask = rootTask ? "task" : "subtask";
-
-      switch (task.status) {
-        case "aborted":
-          return this.$t("task." + taskOrSubtask + "_failed", {
-            action: taskAction,
-          });
-        case "validation-failed":
-          return this.$t("task." + taskOrSubtask + "_failed_validation", {
-            action: taskAction,
-          });
-        case "completed":
-          return this.$t("task." + taskOrSubtask + "_completed", {
-            action: taskAction,
-          });
-        case "pending":
-          return this.$t("task." + taskOrSubtask + "_pending", {
-            action: taskAction,
-          });
-        default:
-          return "";
-      }
-    },
-    createTaskErrorNotification(err, message) {
-      console.error(err);
-
-      const notification = {
-        title: message,
-        description: this.getErrorMessage(err),
-        type: "error",
-      };
-      this.createNotification(notification);
-    },
-  },
-};
